Leaves the database with the ScheduleTypes as they are if there are already some schedule types stored.

+++ Shiftify/shiftify.data/src/main/java/cz/cvut/fit/shiftify/data/DataSeeder.java
@@ -67,6 +67,8 @@ public class DataSeeder {
     private static void initScheduleTypesAndScheduleShifts() {
         ScheduleTypeManager scheduleTypeManager = new ScheduleTypeManager();
         try {
+            if (scheduleTypeManager.scheduleTypes().size() > 0)
+                return;
             scheduleTypeManager.deleteAll();
         } catch (Exception e) {
             e.printStackTrace();
